// frontend/src/app/modules/account/shared/interfaces/account.model.ts
export interface AccountModel {
  activated?: boolean;
  authorities?: string[];
  createdBy?: string;
  createdDate?: string;
  email?: string;
  firstName?: string;
  id?: number;
  imageUrl?: string;
  langKey?: string;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  lastName?: string;
  login?: string;
  color?: string;
  password?: string;
}